requirements.txt
tomli>=2.0.2;python_version<"3.11"
GitPython>=3.1.41
PyGithub>=2.5.0
claude-agent-sdk>=0.1.35
